+++ CHANGELOG.md
@@ -1,3 +1,8 @@
 #Changelog
 
-31.8.2022 - v1.0.0 - initial script for getting data and write to file
+31.8.2022 - v1.0.0 
+- initial script for getting data and write to file
+
+
+03.09.2022 - v1.1.0
+- added the functionality to write into influxDB